backend-services/get_appointments/requirements.txt
functions-framework==3.*
psycopg2-binary
firebase-admin